Further fixes to images
Code no longer errors out when no images provided. Images uploaded are not deleted until new images are uploaded. This is still a bug

+++ ajaxfile.php
@@ -4,21 +4,18 @@ $countFiles = count($_FILES['files']['name']);
 
 
 $dir = "uploads/" . session_id();
-$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
-$files = new RecursiveIteratorIterator($it,
-             RecursiveIteratorIterator::CHILD_FIRST);
-foreach($files as $file) {
-    if ($file->isDir()){
-        rmdir($file->getRealPath());
-    } else {
-        unlink($file->getRealPath());
-    }
+if(is_dir($dir)){
+	$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
+	$files = new RecursiveIteratorIterator($it,RecursiveIteratorIterator::CHILD_FIRST);
+	foreach($files as $file) {
+		if ($file->isDir()){
+			rmdir($file->getRealPath());
+		} else {
+			unlink($file->getRealPath());
+		}
+	}
+	rmdir($dir);
 }
-//rmdir($dir);
-
-
-
-rmdir("uploads/" . session_id());
